Begin refactor of generic inference-job
