--- crates/_types/reusable_types/src/db/mod.rs
@@ -1,1 +1,0 @@
-pub mod payloads;
