--- crates/_types/reusable_types/src/db/payloads/mod.rs
@@ -1,3 +1,0 @@
-/// NB: All fields should be *optional*!
-pub struct MediaUploadDetails {
-}
