+++ crates/lib/database_queries/src/payloads/media_upload_details.rs
@@ -1,0 +1,4 @@
+
+pub struct MediaUploadDetails {
+  // TODO
+}
