+++ crates/service/job/inference_job/src/job_steps/mod.rs
@@ -1,7 +1,6 @@
-pub mod download_file_from_bucket;
-pub mod health_check_trap;
-pub mod job_args;
+pub mod job_dependencies;
 pub mod job_stats;
+pub mod main_loop;
 pub mod process_single_job;
 pub mod process_single_job_error;
-pub mod seconds_to_decoder_steps;
+pub mod sidecar_health_check_trap;
