+++ crates/service/job/inference_job/src/job_steps/sidecar_health_check_trap.rs
