+++ crates/service/job/inference_job/src/job_types/mod.rs
@@ -1,0 +1,2 @@
+pub mod tts;
+pub mod vc;
