+++ crates/service/job/inference_job/src/util/mod.rs
@@ -1,1 +1,2 @@
 pub mod scoped_temp_dir_creator;
+pub mod download_file_from_bucket;
